Popravljen linking bug, promijenjeni intovi u for loopu u size_t jer je i n_ size_t, takodjer dodan typecast u metodi at() za n_ u int jer je n_ size_t a size_t ne moze bit negativan

+++ zad1/MojNizInt.hpp
@@ -23,7 +23,7 @@ class MojNizInt {
     int& operator[](const size_t& i) const {return p_[i];};
 
     //MojNizInt operator*(int&&); //ovo slobodno ignorisi
-    MojNizInt& operator*=(const int& i) {for(int j =0; j<n_; j++) p_[j]*=i; return *this;};
+    MojNizInt& operator*=(const int& i) {for(size_t j =0; j<n_; j++) p_[j]*=i; return *this;};
     MojNizInt& operator+=(const MojNizInt&);
 
     MojNizInt operator++(int); //postfix++ il ti ga sufix
